Fix tick count bug

+++ VecMultiDenseReal128/RunRules/Simulation/include/VecMultiDenseReal.h
@@ -28,9 +28,9 @@ extern "C" {
  * \param [out] outstream_result The stream should be of size (param_m * 4) bytes.
  */
 void VecMultiDenseReal(
-	uint32_t param_m,
-	uint32_t param_n,
-	uint32_t param_original_m,
+	uint64_t param_m,
+	uint64_t param_n,
+	uint64_t param_original_m,
 	const float *instream_coefficients,
 	const float *instream_padded_xs,
 	const float *instream_xs,
@@ -54,9 +54,9 @@ void VecMultiDenseReal(
  * \return A handle on the execution status, or NULL in case of error.
  */
 max_run_t *VecMultiDenseReal_nonblock(
-	uint32_t param_m,
-	uint32_t param_n,
-	uint32_t param_original_m,
+	uint64_t param_m,
+	uint64_t param_n,
+	uint64_t param_original_m,
 	const float *instream_coefficients,
 	const float *instream_padded_xs,
 	const float *instream_xs,
@@ -67,9 +67,9 @@ max_run_t *VecMultiDenseReal_nonblock(
  * 
  */
 typedef struct { 
-	uint32_t param_m; /**<  [in] Interface Parameter "m". */
-	uint32_t param_n; /**<  [in] Interface Parameter "n". */
-	uint32_t param_original_m; /**<  [in] Interface Parameter "original_m". */
+	uint64_t param_m; /**<  [in] Interface Parameter "m". */
+	uint64_t param_n; /**<  [in] Interface Parameter "n". */
+	uint64_t param_original_m; /**<  [in] Interface Parameter "original_m". */
 	const float *instream_coefficients; /**<  [in] The stream should be of size (param_n * 4) bytes. */
 	const float *instream_padded_xs; /**<  [in] The stream should be of size 12288 bytes. */
 	const float *instream_xs; /**<  [in] The stream should be of size (param_m * 4) bytes. */
