+++ VecMultiSparseComplex8/RunRules/Simulation/include/VecMultiSparseComplex.h
@@ -24,14 +24,14 @@ extern "C" {
  * \param [in] param_original_m Interface Parameter "original_m".
  * \param [in] instream_coefficients The stream should be of size ((param_n * 2) * 4) bytes.
  * \param [in] instream_exponents The stream should be of size ((param_n * 2) * 4) bytes.
- * \param [in] instream_padded_xs The stream should be of size 2048 bytes.
+ * \param [in] instream_padded_xs The stream should be of size 1024 bytes.
  * \param [in] instream_xs The stream should be of size ((param_m * 2) * 4) bytes.
  * \param [out] outstream_result The stream should be of size ((param_m * 2) * 4) bytes.
  */
 void VecMultiSparseComplex(
-	uint32_t param_m,
-	uint32_t param_n,
-	uint32_t param_original_m,
+	uint64_t param_m,
+	uint64_t param_n,
+	uint64_t param_original_m,
 	const float *instream_coefficients,
 	const uint32_t *instream_exponents,
 	const float *instream_padded_xs,
@@ -51,15 +51,15 @@ void VecMultiSparseComplex(
  * \param [in] param_original_m Interface Parameter "original_m".
  * \param [in] instream_coefficients The stream should be of size ((param_n * 2) * 4) bytes.
  * \param [in] instream_exponents The stream should be of size ((param_n * 2) * 4) bytes.
- * \param [in] instream_padded_xs The stream should be of size 2048 bytes.
+ * \param [in] instream_padded_xs The stream should be of size 1024 bytes.
  * \param [in] instream_xs The stream should be of size ((param_m * 2) * 4) bytes.
  * \param [out] outstream_result The stream should be of size ((param_m * 2) * 4) bytes.
  * \return A handle on the execution status, or NULL in case of error.
  */
 max_run_t *VecMultiSparseComplex_nonblock(
-	uint32_t param_m,
-	uint32_t param_n,
-	uint32_t param_original_m,
+	uint64_t param_m,
+	uint64_t param_n,
+	uint64_t param_original_m,
 	const float *instream_coefficients,
 	const uint32_t *instream_exponents,
 	const float *instream_padded_xs,
@@ -71,12 +71,12 @@ max_run_t *VecMultiSparseComplex_nonblock(
  * 
  */
 typedef struct { 
-	uint32_t param_m; /**<  [in] Interface Parameter "m". */
-	uint32_t param_n; /**<  [in] Interface Parameter "n". */
-	uint32_t param_original_m; /**<  [in] Interface Parameter "original_m". */
+	uint64_t param_m; /**<  [in] Interface Parameter "m". */
+	uint64_t param_n; /**<  [in] Interface Parameter "n". */
+	uint64_t param_original_m; /**<  [in] Interface Parameter "original_m". */
 	const float *instream_coefficients; /**<  [in] The stream should be of size ((param_n * 2) * 4) bytes. */
 	const uint32_t *instream_exponents; /**<  [in] The stream should be of size ((param_n * 2) * 4) bytes. */
-	const float *instream_padded_xs; /**<  [in] The stream should be of size 2048 bytes. */
+	const float *instream_padded_xs; /**<  [in] The stream should be of size 1024 bytes. */
 	const float *instream_xs; /**<  [in] The stream should be of size ((param_m * 2) * 4) bytes. */
 	float *outstream_result; /**<  [out] The stream should be of size ((param_m * 2) * 4) bytes. */
 } VecMultiSparseComplex_actions_t;
